+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_compose_registries.py
@@ -12,7 +12,7 @@ from azext_containerapp.tests.latest.common import (
     write_test_file,
     clean_up_test_file,
     TEST_DIR, TEST_LOCATION)
-from .utils import create_containerapp_env
+from .utils import create_containerapp_env, prepare_containerapp_env_for_app_e2e_tests
 
 
 class ContainerappComposePreviewRegistryAllArgsScenarioTest(ContainerappComposePreviewScenarioTest):
@@ -20,26 +20,24 @@ class ContainerappComposePreviewRegistryAllArgsScenarioTest(ContainerappComposeP
     @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
     def test_containerapp_compose_create_with_registry_all_args(self, resource_group):
         self.cmd('configure --defaults location={}'.format(TEST_LOCATION))
-
-        compose_text = """
+        app = self.create_random_name(prefix='composewithreg', length=24)
+        compose_text = f"""
 services:
-  foo:
+  {app}:
     image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
     ports: 8080:80
 """
         compose_file_name = f"{self._testMethodName}_compose.yml"
         write_test_file(compose_file_name, compose_text)
-        env_name = self.create_random_name(prefix='containerapp-compose', length=24)
+        env_id = prepare_containerapp_env_for_app_e2e_tests(self)
 
         self.kwargs.update({
-            'environment': env_name,
+            'environment': env_id,
             'compose': compose_file_name,
             'registry_server': "foobar.azurecr.io",
             'registry_user': "foobar",
             'registry_pass': "snafu",
         })
-
-        create_containerapp_env(self, env_name, resource_group)
 
         command_string = 'containerapp compose create'
         command_string += ' --compose-file-path {compose}'
@@ -50,9 +48,9 @@ services:
         command_string += ' --registry-password {registry_pass}'
 
         self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.configuration.registries[0].server', ["foobar.azurecr.io"]),
-            self.check('[?name==`foo`].properties.configuration.registries[0].username', ["foobar"]),
-            self.check('[?name==`foo`].properties.configuration.registries[0].passwordSecretRef', ["foobarazurecrio-foobar"]),  # pylint: disable=C0301
+            self.check(f'[?name==`{app}`].properties.configuration.registries[0].server', ["foobar.azurecr.io"]),
+            self.check(f'[?name==`{app}`].properties.configuration.registries[0].username', ["foobar"]),
+            self.check(f'[?name==`{app}`].properties.configuration.registries[0].passwordSecretRef', ["foobarazurecrio-foobar"]),  # pylint: disable=C0301
         ])
 
         clean_up_test_file(compose_file_name)
@@ -63,24 +61,22 @@ class ContainerappComposePreviewRegistryServerArgOnlyScenarioTest(ContainerappCo
     @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
     def test_containerapp_compose_create_with_registry_server_arg_only(self, resource_group):
         self.cmd('configure --defaults location={}'.format(TEST_LOCATION))
-
-        compose_text = """
+        app = self.create_random_name(prefix='composewithreg', length=24)
+        compose_text = f"""
 services:
-  foo:
+  {app}:
     image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
     ports: 8080:80
 """
         compose_file_name = f"{self._testMethodName}_compose.yml"
         write_test_file(compose_file_name, compose_text)
-        env_name = self.create_random_name(prefix='containerapp-compose', length=24)
+        env_id = prepare_containerapp_env_for_app_e2e_tests(self)
 
         self.kwargs.update({
-            'environment': env_name,
+            'environment': env_id,
             'compose': compose_file_name,
             'registry_server': "foobar.azurecr.io",
         })
-
-        create_containerapp_env(self, env_name, resource_group)
         
         command_string = 'containerapp compose create'
         command_string += ' --compose-file-path {compose}'